requirements.txt
django==2.2.5
pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pytils==0.3
django-solo==1.1.3
django-ckeditor==5.7.1
Fabric==2.5.0
fabric3==1.14.post1
python-dateutil==2.6.0